Fake source IP address for advertisements
Some OS are picky with the source IP address used when receiving
neighbor advertisements. For example, OpenBSD will drop such an
advertisement if the source address is not part of its local subnet.

When sending a neighbor advertisement, spoof the source IPv6 address
to match the target address. This requires a recent kernel (4.15+)
since the socket needs the `IPV6_FREEBIND` option to be set. We cannot
use `IP_FREEBIND` because the socket is of type `SOCK_RAW`.

Fix #41